feat: support self hosted deployment
